refactor(skill): keep audit within complexity budget
