+++ .buildchain/kfd/kfd-3/collaboration-interface.prebuild.json
@@ -6,7 +6,7 @@
   "supportLevel": "release",
   "source": {
     "cwd": ".",
-    "sourceSha": "9d7a0621735da22bbd9b466c077e65f5514200d7",
+    "sourceSha": "186be460e235e77791e4d917aea644afe9d8e28a",
     "registryPath": ".buildchain/kfd/kfd-3/surfaces.json",
     "registryDigest": "sha256:2a978e66d50856462776b86430eebb7c9a5523e8be70120a7201e6dc9df8c7ac"
   },
