+++ developer/sdk/kfd/kfd-2/release-claims.json
@@ -11,7 +11,7 @@
     "version": "4.0.0-alpha.1",
     "channel": "repository-default",
     "tag": "v4.0.0-alpha.1",
-    "sourceSha": "9d7a0621735da22bbd9b466c077e65f5514200d7"
+    "sourceSha": "186be460e235e77791e4d917aea644afe9d8e28a"
   },
   "claims": [
     {
