+++ framework/core/src/python/kungfu/skill/__init__.py
@@ -25,12 +25,16 @@ from .dependencies import (
 )
 from .audit import (
     append_audit_event,
+    build_skill_runtime_audit,
+    project_skill_runtime_audit,
     read_audit_file,
+    read_skill_runtime_audit,
     skill_advertised_event,
     skill_audit_document,
     skill_dependencies_bound_event,
     skill_loaded_event,
     write_audit_document,
+    write_skill_runtime_audit,
 )
 from .parser import SkillError, parse_skill
 from .provider import (
@@ -41,12 +45,6 @@ from .provider import (
     has_context_envelope_info,
     inject_skill_context,
     load_skill_context_file,
-)
-from .runtime_audit import (
-    build_skill_runtime_audit,
-    project_skill_runtime_audit,
-    read_skill_runtime_audit,
-    write_skill_runtime_audit,
 )
 from .registry import (
     SkillRegistryError,
